% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessSnvIndel.R
\name{annotateVariantType}
\alias{annotateVariantType}
\title{Annotate variant type with snpEff}
\usage{
annotateVariantType(
  vcf.file,
  out.file,
  genome = "GRCh37.75",
  java.path = JAVA_PATH,
  snpeff.path = SNPEFF_PATH
)
}
\arguments{
\item{vcf.file}{Path to vcf file (gzip compressed)}

\item{out.file}{Path to output txt file (make sure to add .gz at the end)}

\item{genome}{Name of the genome to supply to snpEff}

\item{java.path}{Path to java binary (defaults to the installed JRE location)}

\item{snpsift.path}{Path to SnpSift jar (defaults to the one included in this package)}
}
\value{
Nothing but writes a gzipped txt file
}
\description{
Annotate variant type with snpEff
}
