% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessDiplotypes.R
\name{detBiallStatus}
\alias{detBiallStatus}
\title{Determine biallelic hit type v2}
\usage{
detBiallStatus(gene.diplotypes)
}
\arguments{
\item{diplotypes}{Diplotypes dataframe}
}
\value{
A character vector of hit types
}
\description{
Determine biallelic hit type v2
}
