% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessSnvIndel.R
\name{detIsHotspotMut}
\alias{detIsHotspotMut}
\title{Determine if mutation is a hotspot mutation}
\usage{
detIsHotspotMut(df, db.path, verbose = T)
}
\arguments{
\item{df}{A bed file like dataframe containing the columns: chrom, pos, ref, alt}

\item{db.path}{Path to a tabix indexed bed file with the clinical significance annotations (e.g.
from ClinVar). This bed file should contain the columns (in this order): chrom, pos, ref, alt, 
clinsig, id. However, it should not have colnames.}
}
\value{
A logical vector
}
\description{
Determine if mutation is a hotspot mutation
}
