% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGeneId.R
\name{ensgToHgncSymbol}
\alias{ensgToHgncSymbol}
\title{Retrieve HGNC symbols from a vector of ENSEMBL gene ids}
\usage{
ensgToHgncSymbol(
  ensembl.gene.ids,
  genes.hgnc = read.delim(GENES_HGNC, stringsAsFactors = F),
  na.rough.fix = F
)
}
\arguments{
\item{ensembl.gene.ids}{A character vector of ENSEMBL gene ids}

\item{genes.hgnc}{A HGNC genes list as a dataframe}

\item{na.rough.fix}{Convert NA values to '' ?}
}
\value{
A character vector of HGNC symbols
}
\description{
This function will attempt to retrieve HGNC gene symbols from a list of gene names
from the HGNC database.
}
