% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessSnvIndel.R
\name{extractVcfFields}
\alias{extractVcfFields}
\title{Extracts revelant fields for downstream functions and outputs as txt file}
\usage{
extractVcfFields(
  vcf.file,
  out.file,
  java.path = JAVA_PATH,
  snpsift.path = SNPSIFT_PATH
)
}
\arguments{
\item{vcf.file}{Path to vcf file (gzip compressed)}

\item{out.file}{Path to output txt file (make sure to add .gz at the end)}

\item{java.path}{Path to java binary (defaults to the installed JRE location)}

\item{snpsift.path}{Path to SnpSift jar (defaults to the one included in this package)}
}
\value{
Nothing but writes a gzipped txt file
}
\description{
Extracts revelant fields for downstream functions and outputs as txt file
}
