% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGeneId.R
\name{geneNamesToEnsg}
\alias{geneNamesToEnsg}
\title{Retrieve ENSEMBL gene ids from a vector of gene names}
\usage{
geneNamesToEnsg(
  gene.names,
  genes.hgnc = read.delim(GENES_HGNC, stringsAsFactors = F),
  na.rough.fix = F
)
}
\arguments{
\item{gene.names}{A character vector of gene names}

\item{genes.hgnc}{A HGNC genes list as a dataframe}

\item{na.rough.fix}{Convert NA values to '' ?}
}
\value{
A character vector of ENSEMBL gene ids
}
\description{
This function will attempt to retrieve ENSEMBL gene ids from a list of gene names
from the HGNC database.
}
