% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessDiplotypes.R
\name{getGeneDiplotypeMaxEff}
\alias{getGeneDiplotypeMaxEff}
\title{Determine the most pathogenic diplotype effect per gene}
\usage{
getGeneDiplotypeMaxEff(gene.diplotypes)
}
\arguments{
\item{gene.diplotypes}{Table containing the mutation info of alleles 1 and 2}

\item{colname}{Which column name to take the max of?}
}
\value{
The subsetted gene.diplotypes table with a column indicating how many rows there were
originally per gene
}
\description{
Determine the most pathogenic diplotype effect per gene
}
