% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkGeneDiplotypes.R
\name{mkGeneDiplotypesCnvMut}
\alias{mkGeneDiplotypesCnvMut}
\title{Merge cnv and germ/som mut profiles into a table the describes the events of allele 1 and 2}
\usage{
mkGeneDiplotypesCnvMut(
  mut.profile.cnv,
  mut.profile.mut,
  mut.origin,
  verbose = T
)
}
\arguments{
\item{mut.profile.cnv}{CNV mut profile table}

\item{mut.profile.mut}{germ or som mut profile table}

\item{mut.origin}{Indicate whether allele 2 originates from 'germ' or 'som'. Used to create
the diplotype_origin column}

\item{verbose}{Print progress messages?}
}
\value{
A table the describes the events of allele 1 and 2
}
\description{
Merge cnv and germ/som mut profiles into a table the describes the events of allele 1 and 2
}
