% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkGeneDiplotypes.R
\name{mkGeneDiplotypesMutMut}
\alias{mkGeneDiplotypesMutMut}
\title{Merge germ/som mut profiles into a table the describes the events of allele 1 and 2}
\usage{
mkGeneDiplotypesMutMut(
  mut.profile.mut1,
  mut.profile.mut2,
  diplotype.origin,
  min.biall.score = c(3, 3),
  get.max.effect.per.gene = T,
  verbose = T
)
}
\arguments{
\item{mut.profile.mut1}{germ/som mut profile table}

\item{mut.profile.mut2}{germ/som mut profile table}

\item{diplotype.origin}{Indicate the mut/mut combination: 'germ_som', 'som_som', 'germ_germ'}

\item{verbose}{Print progress messages?}
}
\value{
A table the describes the events of allele 1 and 2
}
\description{
Per gene, only the most pathogenic germ/som variant pairs are merged. This
prevents the table from growing exponentially.
}
