% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkMutProfile.R
\name{mkMutProfileGeneCnv}
\alias{mkMutProfileGeneCnv}
\title{Determine gene copy number gains and losses}
\usage{
mkMutProfileGeneCnv(
  cnv.file,
  sel.cols = c(chrom = "chromosome", start = "start", end = "end", total_cn =
    "copyNumber", major_cn = "majorAllelePloidy", minor_cn = "minorAllelePloidy"),
  exons.bed.file = EXONS_BED_FILE,
  genes.bed.file = NULL,
  min.gain.ratio.genome = 1.8,
  min.gain.ratio.max.arm.diff = 0.8,
  deep.del.max.max.copy.number = 0.3,
  loh.max.min.minor.allele.ploidy = 0.2,
  verbose = T
)
}
\arguments{
\item{cnv.file}{Path to the cnv bed file.}

\item{sel.cols}{A named vector indicating the columns to keep. Destination colnames should be:
chrom, start, end, total_cn, major_cn, minor_cn}

\item{exons.bed.file}{Path to the bed file containing the columns chrom, exon_start, exon_end, 
hgnc_symbol, ensembl_gene_id}

\item{genes.bed.file}{Alternative to exons.bed.file.}

\item{min.gain.ratio.genome}{The min genome amp ratio (min_copy_number / genome_ploidy) to be 
considered an amplification.}

\item{min.gain.ratio.max.arm.diff}{The min difference between gain_ratio_max and gain_ratio_arm
to be considered a focal amplification}

\item{deep.del.max.max.copy.number}{The max max_copy_number for a gene to be considered to 
be completely lost (deep deletion)}

\item{loh.max.min.minor.allele.ploidy}{The maximum min_minor_allele_ploidy for a gene to be
considered to have loss of heterozygosity}

\item{trunc.max.min.copy.number}{The max min_copy_number to for a gene to be considered to have
a truncation}
}
\value{
The original input dataframe with annotation columns
}
\description{
Annotations deep deletions, truncations, LOH, and amplifications
}
