% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkMutProfile.R
\name{mkMutProfileSnvIndel}
\alias{mkMutProfileSnvIndel}
\title{Create annotations to the germline/somatic tables originating from HMF germline/somatic vcfs}
\usage{
mkMutProfileSnvIndel(
  df.snv.indel,
  scoring = SCORING_MUT,
  keep.only.first.eff = T,
  filter.no.impact.variants = T,
  verbose = T
)
}
\arguments{
\item{df.snv.indel}{A germline/somatic table originating from HMF germline/somatic vcfs}

\item{scoring}{A list containing the scoring for snpeff, clinvar, and enigma annotation values}

\item{keep.only.first.eff}{Only keep first snpeff_eff (items are separated by '&')?}

\item{verbose}{Show messages?}
}
\value{
The original input dataframe with annotation columns
}
\description{
Primarily used to determine
}
