% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detGeneStatuses.R
\name{detGeneStatuses}
\alias{detGeneStatuses}
\title{Determine gene amplifications and biallelic losses}
\usage{
detGeneStatuses(
  out.dir,
  input.file.paths = c(germ_vcf = NA, som_vcf = NA, cnv = NA),
  sample.name,
  genes.bed.file = GENES_BED_FILE,
  exons.bed.file = EXONS_BED_FILE,
  java.path = JAVA_PATH,
  snpsift.path = SNPSIFT_PATH,
  snpeff.path = SNPEFF_PATH,
  do.filter.vcf = T,
  do.snpeff.ann = F,
  keep.chroms = c(1:22, "X"),
  sel.cols.cnv = c(chrom = "chromosome", start = "start", end = "end", total_cn =
    "copyNumber", major_cn = "majorAllelePloidy", minor_cn = "minorAllelePloidy"),
  verbose = T
)
}
\arguments{
\item{out.dir}{Path to output dir}

\item{input.file.paths}{A list or vector supplying the path output files from the HMF pipeline.}

\item{sample.name}{Name of the sample
This list should be in the form c(germ_vcf='', som_vcf='', cnv='')}

\item{genes.bed.file}{Path to the bed file containing the genes of interest. This bed file should 
also contain the column ensembl_gene_id}

\item{exons.bed.file}{Path to the bed file containing the exons from the genes of interest.}

\item{java.path}{Path the the java binary}

\item{snpsift.path}{Path to the SnpSift.jar}

\item{snpeff.path}{Path to snpEff.jar}

\item{do.snpeff.ann}{Annotate SNV/indels variant type with snpeff?
calcChromArmPloidies().}

\item{verbose}{Show progress messages?}
}
\value{
Writes a diplotypes table to the output dir
}
\description{
Determine gene amplifications and biallelic losses
}
